fix auto scale

+++ include/data.h
@@ -6,7 +6,7 @@
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/21 21:24:03 by reasuke           #+#    #+#             */
-/*   Updated: 2024/06/26 20:39:24 by reasuke          ###   ########.fr       */
+/*   Updated: 2024/06/26 20:43:02 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,8 +36,8 @@ typedef struct s_data
 	t_point			**points;
 	int				rows;
 	int				cols;
-	int				orig_z_min;
-	int				orig_z_max;
+	double			orig_z_min;
+	double			orig_z_max;
 }	t_data;
 
 typedef t_point	(*t_operation)(t_point p, double param);
